tweak

+++ 2025/Six/Six.java
@@ -3,8 +3,8 @@ package Six;
 import java.util.ArrayList;
 
 public class Six {
-    public static int Work(ArrayList<String> input){
-       int result = 0;
+    public static long Work(ArrayList<String> input){
+       long result = 0;
        
        
        return result;
