+++ 2025/mal/Mal.java
@@ -3,8 +3,8 @@ package Mal;
 import java.util.ArrayList;
 
 public class Mal {
-    public static int Work(ArrayList<String> input){
-       int result = 0;
+    public static long Work(ArrayList<String> input){
+       long result = 0;
        
        
        return result;
